% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_functions.R
\name{detect_and_convert_to_factors}
\alias{detect_and_convert_to_factors}
\title{Detect and Convert Categorical Variables to Factors}
\usage{
detect_and_convert_to_factors(df, threshold = 10)
}
\arguments{
\item{df}{A data frame.}

\item{threshold}{An integer, the maximum number of unique values for a column
to be considered categorical and converted to a factor.}
}
\value{
A data frame with identified categorical columns converted to factors.
}
\description{
Dynamically detects categorical variables based on a unique value threshold
and converts them to factors.
}
