% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_functions.R
\name{install_and_load_packages}
\alias{install_and_load_packages}
\title{Install and Load Required Packages}
\usage{
install_and_load_packages()
}
\description{
Checks if necessary packages are installed and loads them. If a package is missing, it attempts to install it.
}
\keyword{internal}
