% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_functions.R
\name{read_survey_data}
\alias{read_survey_data}
\title{Read Survey Data from an Excel File}
\usage{
read_survey_data(file_path)
}
\arguments{
\item{file_path}{A string, the path to the Excel file.}
}
\value{
A data frame containing the raw survey data.
}
\description{
Reads survey data from a specified Excel file into a data frame.
}
