% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_functions.R
\name{save_cleaned_data_as_spss}
\alias{save_cleaned_data_as_spss}
\title{Save Cleaned Data as SPSS File}
\usage{
save_cleaned_data_as_spss(df, output_file_path)
}
\arguments{
\item{df}{A data frame, the cleaned data.}

\item{output_file_path}{A string, the path where the SPSS file will be saved.}
}
\description{
Saves the cleaned data frame to an SPSS (.sav) file, preserving variable labels.
}
